+++ docs/CHANGELOG.md
@@ -4,11 +4,29 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v0.2.0] - 14.09.2020
+
+### Added
+
+#### Device API
+
+- Check for firmware updates
+- Start firmware updates
+- Start WPS
+
+### Fixed
+
+- Port from mDNS query is now used
+- Get network overview now also works synchroniously
+- Sopping identify device now also works synchroniously
+- Set user device name now also works synchroniously
+
 ## [v0.1.0] - 28.08.2020
 
 ### Added
 
 #### Device API
+
 - Get LED settings
 - Set LED settings
 - Get connected wifi clients
@@ -18,6 +36,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Get details about master wifi (repeater only)
 
 #### PLC API
+
 - Get details about your powerline network
 - Start and stop identifying your PLC device
 - Rename your device